Move main file into lib_node_client example. Add email address to copyright. Delete files unrelated to lib_node_client.
