+++ lib_node_client/objects_definition.h
@@ -1,11 +1,12 @@
 /**
+ *  Copyright (c) 2024
+ * 
  *  @file object_definition.h
  *  @brief This header file contain one function that dynamically create every object and resource and stores
  *         it in a vector that is returned. This vector is the used when constructing a NodeClient.
  *
  *  @author Bastien Pillonel
  *
- *  @date 6/21/2024
  */
 
 #ifndef OBJECTS_DEFINITION_H
@@ -58,10 +59,10 @@
 #define PHOTOCELL_OBJECT_ID 3419
 #define LED_COLOR_OBJECT_ID 3420
 
-/**
- * @brief Instanciate every NodeObject according to uCIFI standard in one vector
- * @return vector<NodeObject *> contains NodeObject instances
-*/
+ /**
+  * @brief Instanciate every NodeObject according to uCIFI standard in one vector
+  * @return vector<NodeObject *> contains NodeObject instances
+ */
 std::vector<NodeObject *> *initializeObjects();
 
 #endif
