+++ lib_node_client/res_callback.h
@@ -1,11 +1,12 @@
 /**
+ *  Copyright (c) 2024
+ * 
  *  @file res_callback.h
  *  @brief This header file contain the declaration of the ResCallback class used in Ressource class. These callbacks let user defined functions to call when
  *  Read/Write/Exec operation are called on a ressource.
  *
- *  @author Bastien Pillonel
+ *  @author Bastien Pillonel <bastien.pillonel@heig-vd.ch>
  *
- *  @date 4/28/2024
  */
 
 #ifndef RES_CALLBACK_H
@@ -20,36 +21,36 @@ template <class T>
 class ResCallback;
 
 /**
- * @brief Parent class for generic child ResCallback class. 
- * 
+ * @brief Parent class for generic child ResCallback class.
+ *
  */
 class ResCallbackBase
 {
 public:
     /**
      * @brief Destroy the Res Callback Base object
-     * 
+     *
      */
     virtual ~ResCallbackBase() {}
 
     /**
      * @brief Copy instance of ResCallBackBase
-     * 
+     *
      * @return ResCallbackBase*
      */
     virtual ResCallbackBase *clone() const = 0;
 
     /**
      * @brief Move ResCallbackBase attribute into another instance
-     * 
-     * @return ResCallbackBase* 
+     *
+     * @return ResCallbackBase*
      */
     virtual ResCallbackBase *move() noexcept = 0;
 };
 
 /**
  * @brief Class registering callback functions. Used in Resource class
- * 
+ *
  * @tparam T type of argument passed to any callback function registered in ResCallback instance
  */
 template <class T>
@@ -62,34 +63,34 @@ public:
 
     /**
      * @brief Construct a new ResCallback object
-     * 
+     *
      */
     ResCallback() : _functions({}) {}
 
     /**
      * @brief Construct a new ResCallback object by copy
-     * 
-     * @param src 
+     *
+     * @param src
      */
     ResCallback(const ResCallback &src) : _functions(src._functions) {}
 
     /**
      * @brief Construct a new ResCallback object by moving
-     * 
-     * @param src 
+     *
+     * @param src
      */
     ResCallback(ResCallback &&src) : _functions(std::move(src._functions)) {}
 
     /**
      * @brief Destroy the ResCallback object
-     * 
+     *
      */
     ~ResCallback() override {}
 
     /**
      * @brief Copy instance of ResCallBack
-     * 
-     * @return ResCallbackBase* 
+     *
+     * @return ResCallbackBase*
      */
     ResCallbackBase *clone() const override
     {
@@ -98,8 +99,8 @@ public:
 
     /**
      * @brief Move ResCallback attribute into another instance
-     * 
-     * @return ResCallbackBase* 
+     *
+     * @return ResCallbackBase*
      */
     ResCallbackBase *move() noexcept override
     {
@@ -108,7 +109,7 @@ public:
 
     /**
      * @brief Add a callback function to the vector of other callback functions registered
-     * 
+     *
      * @param f function to register
      * @return FunctionPtr shared pointer on the function registered
      */
@@ -121,7 +122,7 @@ public:
 
     /**
      * @brief Remove a callback function registered by searching it's address in the vector
-     * 
+     *
      * @param fp shared pointer previously retrieved from returned value of AddListener function
      */
     void RemoveListener(FunctionPtr fp)
@@ -131,7 +132,7 @@ public:
 
     /**
      * @brief Overload of operator (), calling every function registered
-     * 
+     *
      * @param value in the Resource class, it's the value of the resource that will be passed
      */
     void operator()(T value) const
