--- mbed-os.lib
@@ -1,1 +1,0 @@
-https://github.com/ARMmbed/mbed-os/#7ae592dabe59e9f26a4ce56190af61fb8b46d10d
